requirements.txt
python-dotenv
pinecone-client[grpc]
google-generativeai